+++ src/main.js
@@ -6,6 +6,7 @@ const error = require("./error");
 const sendNotification = require("./sendNotification");
 const getDateTime = require("./getDateTime");
 const { initDirs } = require("./files");
+const updateCheck = require("./updateCheck");
 
 const packageJSON = require("../package.json");
 const settings = require("./settings");
@@ -25,6 +26,8 @@ async function init()
             await initDirs();
 
             await server.init();
+
+            await updateCheck.init();
         }
         catch(err)
         {